pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["FazreilieAlep <mrparksaereilee@gmail.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
psycopg = "^3.1.19"
django-split-settings = "^1.3.1"
pyyaml = "^6.0.1"
pre-commit = "^3.7.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
